pyproject.toml
[tool.poetry]
name = "support-agent-one"
version = "0.1.0"
description = "Support Agent with document search and customer history"
authors = ["Tony Krueger"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pydantic-ai = ">=0.0.52,<0.0.53"
logfire = "3.12.0"
fastapi = ">=0.115.12,<0.116.0"
openai = ">=1.70.0,<2.0.0"
supabase = ">=2.15.0,<3.0.0"
langchain = ">=0.3.23,<0.4.0"
pypdf2 = ">=3.0.1,<4.0.0"
pdfminer-six = ">=20250327,<20250328"
beautifulsoup4 = ">=4.13.3,<5.0.0"
nltk = ">=3.9.1,<4.0.0"
spacy = ">=3.8.5,<4.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
uvicorn = {extras = ["standard"], version = ">=0.34.0,<0.35.0"}
rich = "^14.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
pytest-cov = "^6.1.1"
httpx = "^0.28.1"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
ruff = "^0.11.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"